feat: added postman example
