+++ src/a06-vulnerable-outdated/index.js
@@ -1,8 +1,6 @@
 import { startServer } from 'owasp-shared'
 import { step6Server } from './server.js'
-
-import { startTargetServer } from './routes/profile/helper.js'
-startTargetServer()
+import { startTargetServer } from './routes/profile/index.js'
 
 const start = async function () {
   const fastify = await step6Server()
@@ -10,3 +8,5 @@ const start = async function () {
 }
 
 start()
+
+startTargetServer()
